+++ src/Config/Input/F4U-1D/joystick/MOZA AB9 FFB Base {71DA6210-432E-11f1-8001-444553540000}.diff.lua
@@ -1,0 +1,29 @@
+local diff = {
+	["axisDiffs"] = {
+		["a2003cdnil"] = {
+			["name"] = "Rudder",
+			["removed"] = {
+				[1] = {
+					["key"] = "JOY_RZ",
+				},
+			},
+		},
+		["a3224cd3"] = {
+			["name"] = "Propeller governor handle",
+			["removed"] = {
+				[1] = {
+					["key"] = "JOY_SLIDER1",
+				},
+			},
+		},
+		["a3236cd3"] = {
+			["name"] = "Throttle Lever",
+			["removed"] = {
+				[1] = {
+					["key"] = "JOY_Z",
+				},
+			},
+		},
+	},
+}
+return diff